Reordering the deploy script for sanity (#1304)
