+++ publish/deployed/local-ovm/config.json
@@ -62,9 +62,6 @@
 	"Liquidations": {
 		"deploy": true
 	},
-	"SupplySchedule": {
-		"deploy": true
-	},
 	"Synthetix": {
 		"deploy": true
 	},
@@ -138,9 +135,6 @@
 		"deploy": true
 	},
 	"EtherWrapper": {
-		"deploy": true
-	},
-	"NativeEtherWrapper": {
 		"deploy": true
 	}
 }
